requirements.txt
playwright==1.54.0
python-dotenv==1.1.1
pytest==8.4.2
requests==2.32.5
tzdata==2025.2

